requirements-local.txt
# Reddit API
praw==7.7.1

# Kafka
kafka-python==2.0.2

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Sentiment Analysis
textblob==0.17.1
vaderSentiment==3.3.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
retry==0.9.2

# Data Validation
jsonschema==4.20.0